Add Overview for FRDC-ML in project docs, remove starter-topic
Added a new Overview section in the Writerside/topics documentation, outlining the purpose and usage of our FRDC-ML machine learning backbone. This will give new contributors a clear initial understanding of the project. Also, removed the starter-topic.md from the d.tree, as it was a placeholder and no longer needed.